implement exceptions and controller company

+++ serverdesafio/src/main/java/com/giovani/serverdesafio/service/user/DeleteUserServiceImpl.java
@@ -1,5 +1,9 @@
 package com.giovani.serverdesafio.service.user;
 
+import java.util.Optional;
+
+import com.giovani.serverdesafio.exceptions.DeleteUserException;
+import com.giovani.serverdesafio.model.User;
 import com.giovani.serverdesafio.repository.UserRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +15,18 @@ public class DeleteUserServiceImpl {
   @Autowired
   private UserRepository userRepository;
 
-  public void delete(Long id){
-    userRepository.deleteById(id);
+  public void delete(Long id) throws DeleteUserException{
+    Optional<User> opUser;
+    try{
+      opUser = userRepository.findById(id);
+      if(opUser.isPresent()){
+        userRepository.deleteById(id);
+      }
+      throw new DeleteUserException("user " + id + " does not exist");
+    }catch(DeleteUserException ex){
+      ex.printStackTrace();
+    } 
+
+    
   }
 }
